+++ lib/easy_table.dart
@@ -1,6 +1,7 @@
 library easy_table;
 
 export 'src/easy_table.dart';
+export 'src/easy_table_model.dart';
 export 'src/easy_table_cell.dart';
 export 'src/easy_table_row_color.dart';
 export 'src/easy_table_cell_builder.dart';