+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0
+
+* `EasyTableModel` to handle rows and columns.
+
 ## 0.2.0
 
 * `EasyTableTheme` widget to applies a theme to descendant widgets.